+++ src/dialogs/MuxDlg.h
@@ -21,6 +21,7 @@
 #include "afxwin.h"
 
 #include "configuration\EncoderOptions.h"
+#include "controls\MyResizeDialog.h"
 #include "controls\MyButton.h"
 #include "controls\MyStatic.h"
 #include "controls\MyEdit.h"
@@ -29,7 +30,7 @@
 #include "controls\MyComboBox.h"
 #include "controls\MySliderCtrl.h"
 
-class CMuxDlg : public CDialogEx
+class CMuxDlg : public CMyResizeDialog
 {
     DECLARE_DYNAMIC(CMuxDlg)
 public: